+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - added `loadProgress()` that returns a `0.0 - 1.0` that indicates current asset loading progress
 - added `kaboom()` option `loadingScreen` where you can turn off the default loading screen
 - added `drawMasked()` and `drawSubtracted()`
+- removed `layers()` in favor of parent game objects (see "layers" demo)
 
 ### v2000.2.6